requirements.txt
pandas
requests
playwright
unidecode
sqlalchemy
psycopg2-binary
python-dotenv
groq